update the migrations according to the updated stub

+++ database/migrations/2014_10_12_100000_create_password_reset_tokens_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -23,6 +22,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('password_reset_tokens');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('password_reset_tokens');
+        }
     }
 };
